+++ packages/ricos-content/web/src/converters/plainText/toPlainText/complexPlainText.ts
@@ -75,5 +75,4 @@ Collapsible List
 Yo yo yo!
 
 
-
 `;